# CloudEZBackend/src/tsoa-auto/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ResUser": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"properties": {
							"description": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"shows_count": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReqUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone_number": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserLoginRole": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"role_name": {
						"type": "string",
						"nullable": true
					},
					"role_description": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ResUserLogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"role": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserLoginRole"
							}
						],
						"nullable": true
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResError": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReqUserLogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResSuccess": {
				"properties": {
					"result": {
						"type": "string"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IdBody": {
				"properties": {
					"instanceId": {
						"type": "string"
					}
				},
				"required": [
					"instanceId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InstanceStateName": {
				"type": "string",
				"enum": [
					"stopping",
					"pending",
					"running",
					"stopped",
					"terminated",
					"shutting-down"
				]
			},
			"CreateBody": {
				"properties": {
					"imageId": {
						"type": "string"
					},
					"instanceType": {
						"type": "string"
					},
					"minCount": {
						"type": "number",
						"format": "double"
					},
					"maxCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"imageId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Api-Token": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Veridius-APIs",
		"version": "0.0.1",
		"description": "Veridius-APIs",
		"contact": {}
	},
	"paths": {
		"/user/registerViewer": {
			"post": {
				"operationId": "SaveUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqUser"
							}
						}
					}
				}
			}
		},
		"/user/registerUploader": {
			"post": {
				"operationId": "SaveUserUploader",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqUser"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "UserLogin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ResUserLogin"
										},
										{
											"$ref": "#/components/schemas/ResError"
										}
									]
								}
							}
						}
					}
				},
				"description": "user login",
				"summary": "user login",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReqUserLogin"
							}
						}
					}
				}
			}
		},
		"/user/{userId}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResSuccess"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/ec2/list": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/ec2/start": {
			"post": {
				"operationId": "Start",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"action": {
											"type": "string"
										},
										"instanceId": {
											"type": "string"
										}
									},
									"required": [
										"action",
										"instanceId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IdBody"
							}
						}
					}
				}
			}
		},
		"/ec2/stop": {
			"post": {
				"operationId": "Stop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"action": {
											"type": "string"
										},
										"instanceId": {
											"type": "string"
										}
									},
									"required": [
										"action",
										"instanceId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IdBody"
							}
						}
					}
				}
			}
		},
		"/ec2/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"state": {
												"$ref": "#/components/schemas/InstanceStateName"
											},
											"instanceId": {
												"type": "string"
											}
										},
										"required": [
											"state",
											"instanceId"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"tagsSorter": "alpha",
	"apisSorter": "alpha",
	"operationsSorter": "method"
}